on work OrderEntity

+++ src/main/kotlin/dev/vstd/clothes_renting/controller/form/RequestMoreProductForm.kt
@@ -1,7 +1,7 @@
 package dev.vstd.clothes_renting.controller.form
 
 data class RequestMoreProductForm(
-    val sellerId: Int,
+    val sellerId: Long,
     val clothId1: String,
     val quantity1: Int,
     val clothId2: String,
